Moved bulk operations into py2neo.bulk package
