+++ docs/bulk/index.rst
@@ -1,0 +1,17 @@
+***************************************
+``py2neo.bulk`` -- Bulk data operations
+***************************************
+
+.. automodule:: py2neo.bulk
+
+
+Bulk Load Operations
+====================
+
+.. autofunction:: create_nodes
+
+.. autofunction:: merge_nodes
+
+.. autofunction:: create_relationships
+
+.. autofunction:: merge_relationships
